+++ lisp/core/compile.el
@@ -22,7 +22,7 @@
 ;; Handle options
 (add-hook 'eask-before-command-hook
           (lambda ()
-            (when (eask-strict-p) (setq byte-compile-error-on-warn t))
+            (when (eask-strict-p)      (setq byte-compile-error-on-warn t))
             (when (= eask-verbosity 4) (setq byte-compile-verbose t))))
 
 (defconst eask-compile-log-buffer-name "*Compile-Log*"
@@ -32,8 +32,49 @@
   "Print `*Compile-Log*' buffer."
   (when (get-buffer eask-compile-log-buffer-name)
     (with-current-buffer eask-compile-log-buffer-name
-      (eask-print-log-buffer)
+      (if (and (eask-clean-p) (eask-strict-p))
+          (eask-error (buffer-string))  ; Exit with error code!
+        (eask-print-log-buffer))
       (eask-msg ""))))
+
+(defun eask--byte-compile-file-external-contetnt (filename cmd)
+  "Extract result after executing byte-compile the FILENAME.
+
+The CMD is the command to start a new Emacs session."
+  (with-temp-buffer
+    (insert (shell-command-to-string cmd))
+    (goto-char (point-min))
+    (search-forward filename nil t)
+    (re-search-forward "[ \t\r\n]" nil t)
+    (let ((line (string-trim (thing-at-point 'line))))
+      (if (and (string-prefix-p "Compiling " line)
+               (or (string-match-p "... skipped" line)
+                   (string-match-p "... done" line)))
+          (delete-region (point-min) (line-end-position 1))
+        (delete-region (point-min) (point))))
+    (when (search-forward "(Total of " nil t)
+      (goto-char (point-max))
+      (delete-region (line-beginning-position -1) (point-max)))
+    (string-trim (buffer-string))))
+
+(defun eask--byte-compile-file-external (filename)
+  "Byte compile FILENAME with clean environment by opening a new Emacs session."
+  (let* ((cmd (split-string eask-invocation "\n" t))
+         (cmd (format "\"%s\""(mapconcat #'identity cmd "\" \"")))
+         (args (eask-args))
+         (argv (cl-remove-if
+                (lambda (arg)
+                  (or (string= "--clean" arg)  ; prevent infinite call
+                      (member arg args)))      ; remove repeated arguments
+                (eask-argv-out)))
+         (args (append `(,(eask-command) ,(concat "\"" filename "\"")) argv))
+         (args (mapconcat #'identity args " "))
+         (cmd (concat cmd " " args))
+         (content (eask--byte-compile-file-external-contetnt filename cmd)))
+    (if (string-empty-p content)
+        t  ; no error, good!
+      (with-current-buffer (get-buffer-create eask-compile-log-buffer-name)
+        (insert content)))))
 
 (defun eask--byte-compile-file (filename)
   "Byte compile FILENAME."
@@ -45,7 +86,9 @@
     (eask-with-progress
       (unless byte-compile-verbose (format "Compiling %s... " filename))
       (eask-with-verbosity 'debug
-        (setq result (byte-compile-file filename)
+        (setq result (if (eask-clean-p)
+                         (eask--byte-compile-file-external filename)
+                       (byte-compile-file filename))
               result (eq result t)))
       (if result "done ✓" "skipped ✗"))
     (eask--print-compile-log)